+++ packages/webpack/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## 0.5.4 (Apr 23, 2018)
 
+* Add `chunksSortMode` option into `html-webpack-plugin`
 * Replace `DefinePlugin` into `DotenvPlugin` at common config